+++ Assets/Scripts/CarPaint.cs
@@ -6,10 +6,10 @@ using UnityEngine;
 public class CarPaint : ScriptableObject
 {
     public string paintName;
-    public int changeSpeed;
-    public int changeHandling;
-    public int changeAcceleration;
-    public int changeCoolness;
+    public float changeSpeed;
+    public float changeHandling;
+    public float changeAcceleration;
+    public float changeCoolness;
 
     public Material material;
 }
